Move main router to App.
No slash at the end
